services/storage/src/requirements.txt
transformers 
sentence-transformers 
faiss-cpu 
torch
pdfplumber
python-docx
docs
bs4